char: sbt_lockamp: Rework generator sysfs interface
This adds support for the new version 19 of the FPGA image. It also breaks
backwards compatbility with the old sysfs interface.

Now, the generators stepping can be finely controlled together with the
corresponding MA and CIC filters.

The sysfs interface is intentionally low-level. That is, it merely
reflects the newly added hardware registers. It is up to the user to use
all said registers in a manner that makes sense to the hardware.

Usually, sysfs interfaces hide such interaction. In this case, however,
the complexity involved would cause too much of a development and maintenance burden.
There were two options:

  1. Expose a simple sysfs interface that hides some of the advanced
     HW features.
  2. Expose the low-level interface and let it be up to the user to
     implement either a simple or advanced interface on top of sysfs.

We choose option 2. This way, the user is not restricted by arbitrary
decisions in the sysfs interface. On the other hand, the user is also
able to set combinations of sysfs parameters in a way that doesn't make
sense.

+++ drivers/char/sbt_lockamp/lockin_amplifier.h
@@ -47,7 +47,7 @@ struct lockamp {
 	dev_t chrdev_no;
 	struct gpio_desc *reset;
 	struct iio_channel *adc_site0, *adc_site1, *dac_site0, *dac_site1;
-	u32 __iomem *control;
+	u8 __iomem *control;
 	struct regulator *amp_supply;
 	bool amp_supply_force_off;
